feat(asset): split into local/remote
